% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{action}
\alias{action}
\title{Action}
\usage{
action(args, id, driver, ...)
}
\arguments{
\item{args}{A list of arguments to pass to the action.}

\item{id}{The ID of the component.}

\item{driver}{The driver object.}

\item{...}{Additional arguments.}
}
\description{
Action
}
