% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{Driver}
\alias{Driver}
\title{Robust App Driver}
\description{
Robust App Driver

Robust App Driver
}
\details{
Instead of using Shiny IDs of components, use `data-testable-id` that can change independently from Shiny IDs.

Easily detect what is the component we're interacting with using `data-testable-type` to dispatch a proper action.
}
\section{Super class}{
\code{\link[shinytest2:AppDriver]{shinytest2::AppDriver}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-dispatch}{\code{Driver$dispatch()}}
\item \href{#method-NA-get_value}{\code{Driver$get_value()}}
\item \href{#method-NA-get}{\code{Driver$get()}}
\item \href{#method-NA-is_visible}{\code{Driver$is_visible()}}
\item \href{#method-NA-is_disabled}{\code{Driver$is_disabled()}}
\item \href{#method-NA-expect_output_errors}{\code{Driver$expect_output_errors()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="click"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-click'><code>shinytest2::AppDriver$click()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="expect_download"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-expect_download'><code>shinytest2::AppDriver$expect_download()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="expect_html"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-expect_html'><code>shinytest2::AppDriver$expect_html()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="expect_js"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-expect_js'><code>shinytest2::AppDriver$expect_js()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="expect_screenshot"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-expect_screenshot'><code>shinytest2::AppDriver$expect_screenshot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="expect_text"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-expect_text'><code>shinytest2::AppDriver$expect_text()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="expect_unique_names"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-expect_unique_names'><code>shinytest2::AppDriver$expect_unique_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="expect_values"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-expect_values'><code>shinytest2::AppDriver$expect_values()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_chromote_session"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_chromote_session'><code>shinytest2::AppDriver$get_chromote_session()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_dir"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_dir'><code>shinytest2::AppDriver$get_dir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_download"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_download'><code>shinytest2::AppDriver$get_download()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_html"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_html'><code>shinytest2::AppDriver$get_html()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_js"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_js'><code>shinytest2::AppDriver$get_js()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_logs"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_logs'><code>shinytest2::AppDriver$get_logs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_screenshot"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_screenshot'><code>shinytest2::AppDriver$get_screenshot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_text"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_text'><code>shinytest2::AppDriver$get_text()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_url"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_url'><code>shinytest2::AppDriver$get_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_values"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_values'><code>shinytest2::AppDriver$get_values()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_variant"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_variant'><code>shinytest2::AppDriver$get_variant()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="get_window_size"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-get_window_size'><code>shinytest2::AppDriver$get_window_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="initialize"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-initialize'><code>shinytest2::AppDriver$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="log_message"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-log_message'><code>shinytest2::AppDriver$log_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="run_js"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-run_js'><code>shinytest2::AppDriver$run_js()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="set_inputs"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-set_inputs'><code>shinytest2::AppDriver$set_inputs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="set_window_size"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-set_window_size'><code>shinytest2::AppDriver$set_window_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="stop"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-stop'><code>shinytest2::AppDriver$stop()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="upload_file"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-upload_file'><code>shinytest2::AppDriver$upload_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="view"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-view'><code>shinytest2::AppDriver$view()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="wait_for_idle"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-wait_for_idle'><code>shinytest2::AppDriver$wait_for_idle()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="wait_for_js"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-wait_for_js'><code>shinytest2::AppDriver$wait_for_js()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinytest2" data-topic="AppDriver" data-id="wait_for_value"><a href='../../shinytest2/html/AppDriver.html#method-AppDriver-wait_for_value'><code>shinytest2::AppDriver$wait_for_value()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-dispatch"></a>}}
\if{latex}{\out{\hypertarget{method-NA-dispatch}{}}}
\subsection{Method \code{dispatch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Driver$dispatch(testid, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{testid}}{character}

\item{\code{...}}{Object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_value}{}}}
\subsection{Method \code{get_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Driver$get_value(
  testid = missing_arg(),
  input = missing_arg(),
  output = missing_arg(),
  export = missing_arg()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{testid}}{character}

\item{\code{input}}{character}

\item{\code{output}}{character}

\item{\code{export}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Driver$get(testid, code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{testid}}{character}

\item{\code{code}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-is_visible"></a>}}
\if{latex}{\out{\hypertarget{method-NA-is_visible}{}}}
\subsection{Method \code{is_visible()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Driver$is_visible(testid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{testid}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-is_disabled"></a>}}
\if{latex}{\out{\hypertarget{method-NA-is_disabled}{}}}
\subsection{Method \code{is_disabled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Driver$is_disabled(testid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{testid}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-expect_output_errors"></a>}}
\if{latex}{\out{\hypertarget{method-NA-expect_output_errors}{}}}
\subsection{Method \code{expect_output_errors()}}{
Fails a test when there are Shiny errors visible on the page
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Driver$expect_output_errors(n = 0)}\if{html}{\out{</div>}}
}

}
}
