% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify_no_output_error}
\alias{verify_no_output_error}
\title{Verify Shiny app output}
\usage{
verify_no_output_error(driver)
}
\arguments{
\item{driver}{A driver object.}
}
\description{
Verify Shiny app output
}
