% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testable_component.R
\name{testable_component}
\alias{testable_component}
\title{Testable component}
\usage{
testable_component(x, id, data_testable_id, data_testable_type = NULL)
}
\arguments{
\item{x}{A widget}

\item{id}{A string, the ID of the widget}

\item{data_testable_id}{A string, the testable ID of the widget}

\item{data_testable_type}{A string, the testable type of the widget}
}
\description{
Testable component
}
